+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.4] - 2025-01-22
+### Added
+- Graph algorithms:
+  - Bellman-Ford Algorithm
+  - Dijkstra's Algorithm
+  - Floyd-Warshall Algorithm
+  - Johnson's Algorithm
+  - Kruskal's Algorithm
+  - Prim's Algorithm
+  - Tarjan's Algorithm (SCC)
+  - Topological Sort
+  - Warshall's Algorithm
+
 ## [0.6.3] - 2025-01-21
 ### Fixed
 - Release workflow now creates release and publishes crate